fixed price and added constructor

+++ src/main/resources/init_db.sql
@@ -2,6 +2,6 @@ USE book;
 CREATE TABLE IF NOT EXISTS book (
     id BIGINT NOT NULL AUTO_INCREMENT,
     title VARCHAR(255) NOT NULL,
-    price BIGINT NOT NULL,
+    price DECIMAL(10,2) NOT NULL,
     PRIMARY KEY (id)
     );
